backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.3
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.7
defusedxml==0.7.1
Django==5.2.8
django-db-logger==0.1.13
django-nested-admin==4.1.6
django-phonenumber-field==8.1.0
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.14
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.13
psycopg-binary==3.2.13
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-monkey-business==1.1.0
python3-openid==3.2.0
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
